+++ rustls/src/crypto/ring/mod.rs
@@ -14,6 +14,9 @@ use core::fmt;
 
 pub(crate) mod hash;
 pub(crate) mod hmac;
+#[cfg(feature = "tls12")]
+pub(crate) mod tls12;
+pub(crate) mod tls13;
 
 /// Default crypto provider.
 #[derive(Debug)]
@@ -29,9 +32,36 @@ impl CryptoProvider for Ring {
     }
 
     fn default_cipher_suites() -> &'static [SupportedCipherSuite] {
-        crate::suites::DEFAULT_CIPHER_SUITES
+        DEFAULT_CIPHER_SUITES
     }
 }
+
+/// The cipher suite configuration that an application should use by default.
+///
+/// This will be [`ALL_CIPHER_SUITES`] sans any supported cipher suites that
+/// shouldn't be enabled by most applications.
+pub static DEFAULT_CIPHER_SUITES: &[SupportedCipherSuite] = ALL_CIPHER_SUITES;
+
+/// A list of all the cipher suites supported by the rustls *ring* provider.
+pub static ALL_CIPHER_SUITES: &[SupportedCipherSuite] = &[
+    // TLS1.3 suites
+    tls13::TLS13_AES_256_GCM_SHA384,
+    tls13::TLS13_AES_128_GCM_SHA256,
+    tls13::TLS13_CHACHA20_POLY1305_SHA256,
+    // TLS1.2 suites
+    #[cfg(feature = "tls12")]
+    tls12::TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,
+    #[cfg(feature = "tls12")]
+    tls12::TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,
+    #[cfg(feature = "tls12")]
+    tls12::TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256,
+    #[cfg(feature = "tls12")]
+    tls12::TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,
+    #[cfg(feature = "tls12")]
+    tls12::TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,
+    #[cfg(feature = "tls12")]
+    tls12::TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256,
+];
 
 /// An in-progress key exchange.  This has the algorithm,
 /// our private key, and our public key.
